+++ BookBridge/BookBridge/Views/BookShelf/BookDetailView.swift
@@ -1,0 +1,21 @@
+//
+//  BookDetailView.swift
+//  BookBridge
+//
+//  Created by 김건호 on 2/1/24.
+//
+
+import SwiftUI
+
+struct BookDetailView: View {
+    var book: Item
+    
+    var body: some View {
+        VStack{
+            
+        }
+    }
+        
+}
+
+
