+++ BookBridge/BookBridge/Views/BookShelf/BookSearchBar.swift
@@ -1,0 +1,40 @@
+//
+//  BookSearchBar.swift
+//  BookBridge
+//
+//  Created by 김건호 on 2/2/24.
+//
+
+import SwiftUI
+
+struct BookSearchBar: View {
+    @Binding var text: String
+    var placeholder: String
+    
+    var body: some View {
+        HStack {
+            TextField("\(placeholder)", text: $text)
+                .padding(7)
+                .padding(.horizontal, 25)
+                .background(Color(.systemGray6))
+                .cornerRadius(8)
+                .overlay(
+                    HStack {
+                        Image(systemName: "magnifyingglass")
+                            .foregroundColor(.gray)
+                            .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
+                            .padding(.leading, 8)
+                        
+                        if text != "" {
+                            Button(action: {
+                                self.text = ""
+                            }) {
+                                Image(systemName: "multiply.circle.fill")
+                                    .foregroundColor(.gray)
+                                    .padding(.trailing, 8)
+                            }
+                        }
+                    })
+        }
+    }
+}
