+++ BookBridge/BookBridge/Views/BookShelf/BookShelfView.swift
@@ -21,6 +21,7 @@ struct BookShelfView: View {
     @State private var selectedPicker: tapInfo = .wish
     @State private var showingSheet = false // 시트 표시 여부를 위한 상태 변수
     @State private var searchText = ""
+    @State private var selectedBook: Item?
     
     var searchBarPlaceholder: String {
         switch selectedPicker {
@@ -45,14 +46,18 @@ struct BookShelfView: View {
                     .frame(height: 20)
                 
                 
-                SearchBar(text: $searchText, placeholder: searchBarPlaceholder)
+                BookSearchBar(text: $searchText, placeholder: searchBarPlaceholder)
                 
                 
                 Spacer()
                     .frame(height: 20)
                 
-                BookView(tap: selectedPicker)
+                BookView(selectedBook: $selectedBook, tap: selectedPicker)
                     .environmentObject(viewModel)
+                    .sheet(item: $selectedBook) { book in
+                                BookDetailView(book: book)
+                                    .navigationTitle("도서정보")
+                            }
                 
             }
             .padding(20)
@@ -68,96 +73,7 @@ struct BookShelfView: View {
     }
 }
 
-struct BookView: View {
-    @EnvironmentObject var viewModel: BookShelfViewModel
-    var tap: tapInfo
-    
-    var books: [Item] {
-        switch tap {
-        case .wish:
-            return viewModel.wishBooks.flatMap { $0.items }
-        case .hold:
-            return viewModel.holdBooks.flatMap { $0.items }
-        }
-    }
-    
-    let columns: [GridItem] = Array(repeating: .init(.flexible()), count: 3)
-    
-    var body: some View {
-        ScrollView {
-            LazyVGrid(columns: columns, spacing: 20) {
-                ForEach(books, id: \.id) { book in
-                    VStack {
-                        if let urlString = book.volumeInfo.imageLinks?.smallThumbnail, let url = URL(string: urlString) {
-                            AsyncImage(url: url)
-                                .frame(width: (UIScreen.main.bounds.width - 60) / 3, height: 164)
-                                .cornerRadius(8)
-                        }
-                        Text(book.volumeInfo.title ?? "Unknown Title")
-                            .frame(maxWidth: .infinity, alignment: .center) // 텍스트 뷰 크기 조정
-                    }
-                    .frame(width: (UIScreen.main.bounds.width - 60) / 3) // VStack에 대한 프레임 설정
-                }
-            }
-        }
-    }
-}
 
 
-struct SearchBar: View {
-    @Binding var text: String
-    var placeholder: String
-    
-    var body: some View {
-        HStack {
-            TextField("\(placeholder)", text: $text)
-                .padding(7)
-                .padding(.horizontal, 25)
-                .background(Color(.systemGray6))
-                .cornerRadius(8)
-                .overlay(
-                    HStack {
-                        Image(systemName: "magnifyingglass")
-                            .foregroundColor(.gray)
-                            .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                            .padding(.leading, 8)
-                        
-                        if text != "" {
-                            Button(action: {
-                                self.text = ""
-                            }) {
-                                Image(systemName: "multiply.circle.fill")
-                                    .foregroundColor(.gray)
-                                    .padding(.trailing, 8)
-                            }
-                        }
-                    })
-        }
-    }
-}
 
-struct AddBookBtnView: View {
-    @Binding var showingSheet: Bool
-    
-    var body: some View {
-        VStack {
-            Spacer()
-            
-            HStack {
-                Spacer()
-                
-                Button(
-                    action: {
-                        showingSheet = true                            
-                    },
-                    label: {
-                        Image(systemName: "plus.circle.fill")
-                            .resizable()
-                            .foregroundColor(Color(hex:"F6A65D"))
-                            .frame(width: 50,height: 50)
-                    }
-                )
-            }
-        }
-    }
-}
+
